Got rid of all 'Images' In player Feedback and updated scripts. I buggered up the Tutorial though (Sorry Colin <3)
I aimed to get rid of all instances that called for images and renamed a few particles to bettter fit what they actually do, placed particles in appropriate positions. I think the tutorial is a little messed up now but it needed a makeover anyway. Once we get the feedback correlated from that